retry, flatmap, and switchToLatest tests
