+++ SwiftUI-NotesTests/CombinePatternTests.swift
@@ -14,10 +14,6 @@ class CombinePatternTests: XCTestCase {
     enum testFailureCondition: Error {
         case invalidServerResponse
     }
-    // matching the data structure returned from ip.jsontest.com
-    struct IPInfo: Codable {
-            var ip: String
-    }
 
     func testDeadSimpleChain() {
         let simplePublisher = PassthroughSubject<String, Error>()
@@ -111,319 +107,7 @@ class CombinePatternTests: XCTestCase {
     }
 
 
-    func testBasicFlatMap_String_NeverPublisher() {
-        // setup
-        let simpleControlledPublisher = PassthroughSubject<String, Never>()
 
-        let _ = simpleControlledPublisher
-            .flatMap { someValue in // takes a String in and returns a Publisher
-                return Just<String>("Alternate data")
-                // flatMap returns a Publisher, where map returns <Input> - String in this case
-            }
-            .eraseToAnyPublisher()
-            .sink(receiveCompletion: { fini in
-                print(".sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(".sink() received \(stringValue)")
-                // this print adds into the console output:
-                // .sink() received Alternate data
-                // .sink() received Alternate data
-                // .sink() received Alternate data
-                // .sink() received Alternate data
-            })
-
-        let oneFish = "onefish"
-        let twoFish = "twofish"
-        let redFish = "redfish"
-        let blueFish = "bluefish"
-
-        simpleControlledPublisher.send(oneFish)
-        simpleControlledPublisher.send(twoFish)
-        simpleControlledPublisher.send(redFish)
-        simpleControlledPublisher.send(blueFish)
-
-    }
-
-    func testBasicFlatMapWithBackdoorPublisher_String_NeverPublisher() {
-        // setup
-        let simpleControlledPublisher = PassthroughSubject<String, Never>()
-
-        let backDoorPublisher = PassthroughSubject<String, Never>()
-
-        let _ = simpleControlledPublisher
-            .flatMap { someValue -> AnyPublisher<String, Never> in // takes a String in and returns a Publisher
-                return backDoorPublisher.eraseToAnyPublisher()
-            }
-            .eraseToAnyPublisher()
-//            .print()
-            .sink(receiveCompletion: { fini in
-                print(" ** .sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(" ** .sink() received \(stringValue)")
-                // this print adds into the console output:
-                // .sink() received Alternate data
-                // .sink() received Alternate data
-                // .sink() received Alternate data
-                // .sink() received Alternate data
-            })
-
-        let oneFish = "onefish"
-        let twoFish = "twofish"
-        let redFish = "redfish"
-        let blueFish = "bluefish"
-
-        simpleControlledPublisher.send(oneFish)
-        backDoorPublisher.send("first response")
-        // backDoorPublisher.send(completion: .finished)
-        // with the above line uncommented, we only receive:
-        //    ** .sink() received first response
-        //    ** .sink() received second response
-        // and the pipeline appears to be terminated is terminated
-
-        simpleControlledPublisher.send(twoFish)
-        backDoorPublisher.send("second response")
-        
-        // simpleControlledPublisher.send(completion: .finished)
-        // with the above line uncommented, the original pipeline is terminated, but the
-        // backDoor pipelines put into place by the flatmap are still completely active to downstream
-        // subscribers. Console output:
-            //** .sink() received first response
-            //** .sink() received second response
-            //** .sink() received second response
-            //** .sink() received third response
-            //** .sink() received third response
-            //** .sink() received fourth response
-            //** .sink() received fourth response
-            //** .sink() received the completion: finished
-
-        simpleControlledPublisher.send(redFish)
-        backDoorPublisher.send("third response")
-        simpleControlledPublisher.send(blueFish)
-        backDoorPublisher.send("fourth response")
-        backDoorPublisher.send(completion: .finished)
-
-        simpleControlledPublisher.send(blueFish)
-        backDoorPublisher.send("fifth response")
-
-        // based on this output, flatMap is adding a publisher for every element in the original stream
-        // and each publisher that's created gets added - so if the original stream had 3 events flow through,
-        // there could be 3 active publishers sending data
-        //** .sink() received first response
-        //** .sink() received second response
-        //** .sink() received second response
-        //** .sink() received third response
-        //** .sink() received third response
-        //** .sink() received third response
-        //** .sink() received fourth response
-        //** .sink() received fourth response
-        //** .sink() received fourth response
-        //** .sink() received fourth response
-
-    }
-
-    func testBasicFlatMapFallback_Data_NeverPublisher() {
-        // setup
-        let simpleControlledPublisher = PassthroughSubject<Data, Never>()
-
-        let _ = simpleControlledPublisher
-            .flatMap { value in // takes a String in and returns a Publisher
-                return Just<Data>(value)
-                .decode(type: IPInfo.self, decoder: JSONDecoder())
-                .catch { _ in
-                    return Just(IPInfo(ip: "8.8.8.8"))
-                }
-            }
-            .sink(receiveCompletion: { fini in
-                print(".sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(".sink() received \(stringValue)")
-                // this print adds into the console output:
-                // .sink() received IPInfo(ip: "1.2.3.4")
-                // .sink() received IPInfo(ip: "192.168.1.1")
-                // .sink() received IPInfo(ip: "8.8.8.8")
-                // .sink() received IPInfo(ip: "192.168.0.1")
-                // .sink() received the completion: finished
-            })
-
-        let oneFish = "{ \"ip\": \"1.2.3.4\" }".data(using: .utf8)
-        let twoFish = "{ \"ip\": \"192.168.1.1\" }".data(using: .utf8)
-        let redFish = "Opps, crap - no JSON here".data(using: .utf8)
-        let blueFish = "{ \"ip\": \"192.168.0.1\" }".data(using: .utf8)
-
-        simpleControlledPublisher.send(oneFish!)
-        simpleControlledPublisher.send(twoFish!)
-        simpleControlledPublisher.send(redFish!)
-        simpleControlledPublisher.send(blueFish!)
-        simpleControlledPublisher.send(completion: Subscribers.Completion.finished)
-
-    }
-
-    func testBasicFlatMapFallback_Data_ErrorPublisher() {
-        // setup
-        let simpleControlledPublisher = PassthroughSubject<Data, Error>()
-
-        let _ = simpleControlledPublisher
-            .flatMap { value in // takes a String in and returns a Publisher
-                return Just(value)
-                .decode(type: IPInfo.self, decoder: JSONDecoder())
-//                .catch { _ in
-//                    return Publishers.Just(IPInfo(ip: "8.8.8.8"))
-//                }
-            }
-            .sink(receiveCompletion: { fini in
-                print(".sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(".sink() received \(stringValue)")
-                // this print adds into the console output:
-                // .sink() received IPInfo(ip: "1.2.3.4")
-                // .sink() received IPInfo(ip: "192.168.1.1")
-                // .sink() received IPInfo(ip: "8.8.8.8")
-                // .sink() received IPInfo(ip: "192.168.0.1")
-                // .sink() received the completion: finished
-            })
-
-        let oneFish = "{ \"ip\": \"1.2.3.4\" }".data(using: .utf8)
-        let twoFish = "{ \"ip\": \"192.168.1.1\" }".data(using: .utf8)
-        let redFish = "Opps, crap - no JSON here".data(using: .utf8)
-        let blueFish = "{ \"ip\": \"192.168.0.1\" }".data(using: .utf8)
-
-        simpleControlledPublisher.send(oneFish!)
-        simpleControlledPublisher.send(twoFish!)
-        simpleControlledPublisher.send(redFish!)
-        simpleControlledPublisher.send(blueFish!)
-        simpleControlledPublisher.send(completion: Subscribers.Completion.finished)
-
-    }
-
-    func testRetryOperatorWithPassthroughSubject() {
-        // setup
-        let simpleControlledPublisher = PassthroughSubject<String, Error>()
-
-        let _ = simpleControlledPublisher
-            .print()
-            .retry(1)
-            .sink(receiveCompletion: { fini in
-                print(" ** .sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(" ** .sink() received \(stringValue)")
-            })
-
-        let oneFish = "onefish"
-        let twoFish = "twofish"
-        let redFish = "redfish"
-        let blueFish = "bluefish"
-
-        simpleControlledPublisher.send(oneFish)
-        simpleControlledPublisher.send(twoFish)
-
-        // with an error response, this prints two results and hangs...
-        simpleControlledPublisher.send(completion: Subscribers.Completion.failure(testFailureCondition.invalidServerResponse))
-
-        // with a completion, this prints two results and ends
-        //simpleControlledPublisher.send(completion: .finished)
-
-        simpleControlledPublisher.send(redFish)
-        simpleControlledPublisher.send(blueFish)
-    }
-
-    func testRetryOperatorWithCurrentValueSubject() {
-        // setup
-        let simpleControlledPublisher = CurrentValueSubject<String, Error>("initial value")
-
-        let _ = simpleControlledPublisher
-            .print("(1)>")
-            .retry(3)
-            .print("(2)>")
-            .sink(receiveCompletion: { fini in
-                print(" ** .sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(" ** .sink() received \(stringValue)")
-            })
-
-        let oneFish = "onefish"
-
-        simpleControlledPublisher.send(oneFish)
-        // with an error response, this prints two results and hangs...
-        simpleControlledPublisher.send(completion: Subscribers.Completion.failure(testFailureCondition.invalidServerResponse))
-
-        // with a completion, this prints two results and ends
-        //simpleControlledPublisher.send(completion: .finished)
-
-//        output:
-//        (1)>: receive subscription: (CurrentValueSubject)
-//        (2)>: receive subscription: (Retry)
-//        (2)>: request unlimited
-//        (1)>: request unlimited
-//        (1)>: receive value: (initial value)
-//        (2)>: receive value: (initial value)
-//        ** .sink() received initial value
-//        (1)>: receive value: (onefish)
-//        (2)>: receive value: (onefish)
-//        ** .sink() received onefish
-//        (1)>: receive finished
-//        (2)>: receive finished
-//        ** .sink() received the completion: finished
-    }
-
-    func testRetryWithOneShotJustPublisher() {
-        // setup
-        let _ = Just<String>("yo")
-            .print("(1)>")
-            .retry(3)
-            .print("(2)>")
-            .sink(receiveCompletion: { fini in
-                print(" ** .sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(" ** .sink() received \(stringValue)")
-            })
-//        output:
-//        (1)>: receive subscription: (Just)
-//        (2)>: receive subscription: (Retry)
-//        (2)>: request unlimited
-//        (1)>: request unlimited
-//        (1)>: receive value: (yo)
-//        (2)>: receive value: (yo)
-//        ** .sink() received yo
-//        (1)>: receive finished
-//        (2)>: receive finished
-//        ** .sink() received the completion: finished
-
-    }
-
-    func testRetryWithOneShotFailPublisher() {
-        // setup
-        let _ = Publishers.Fail(outputType: String.self, failure: testFailureCondition.invalidServerResponse)
-            .print("(1)>")
-            .retry(3)
-            .print("(2)>")
-            .sink(receiveCompletion: { fini in
-                print(" ** .sink() received the completion:", String(describing: fini))
-            }, receiveValue: { stringValue in
-                XCTAssertNotNil(stringValue)
-                print(" ** .sink() received \(stringValue)")
-            })
-//        output:
-//        (1)>: receive subscription: (Empty)
-//        (1)>: receive error: (invalidServerResponse)
-//        (1)>: receive subscription: (Empty)
-//        (1)>: receive error: (invalidServerResponse)
-//        (1)>: receive subscription: (Empty)
-//        (1)>: receive error: (invalidServerResponse)
-//        (1)>: receive subscription: (Empty)
-//        (1)>: receive error: (invalidServerResponse)
-//        (2)>: receive error: (invalidServerResponse)
-//        ** .sink() received the completion: failure(SwiftUI_NotesTests.CombinePatternTests.testFailureCondition.invalidServerResponse)
-//        (2)>: receive subscription: (Retry)
-//        (2)>: request unlimited
-
-    }
 
 
     func testFutureSignatureWithoutErasure() {
